Feat: exit option component #51
