+++ frontend/src/components/Options/Account/index.js
@@ -1,0 +1,1 @@
+export {default as ExitOption} from './exitOption.js';
